requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mimesis==4.1.3
openapi-codec==1.3.2
Pillow==8.1.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
